Add basic table for statistics
