+++ src/gui/controls.rs
@@ -1,9 +1,7 @@
 //! GUI code for all the separate controls (buttons, text_input, checkboxes, etc.)
 
-use std::thread::{self, available_parallelism};
-
 use cadical::Solver;
-use egui::{vec2, FontId, Key, Label, Response, RichText, ScrollArea, TextStyle, Ui};
+use egui::{vec2, FontId, Key, Label, Response, RichText, TextStyle, Ui};
 
 use super::SATApp;
 
@@ -290,306 +288,6 @@ impl SATApp {
                 RichText::new("Trail with conflict literals").size(text_scale),
             ));
         });
-    }
-
-    /// Show statistics
-    fn statistics(&mut self, ui: &mut Ui, ctx: &egui::Context, text_scale: f32) {
-        ui.horizontal(|ui| {
-            if ui
-                .button(RichText::new("Statistics").size(text_scale))
-                .clicked()
-            {
-                self.state.show_statistics = true;
-            }
-
-            if ui
-                .button(RichText::new("Process with all configurations").size(text_scale))
-                .clicked()
-            {
-                let encodings = vec![
-                    EncodingType::Decimal {
-                        cell_at_least_one: true,
-                        cell_at_most_one: false,
-                        sudoku_has_all_values: false,
-                        sudoku_has_unique_values: true,
-                    },
-                    EncodingType::Decimal {
-                        cell_at_least_one: true,
-                        cell_at_most_one: true,
-                        sudoku_has_all_values: false,
-                        sudoku_has_unique_values: true,
-                    },
-                    EncodingType::Decimal {
-                        cell_at_least_one: true,
-                        cell_at_most_one: false,
-                        sudoku_has_all_values: true,
-                        sudoku_has_unique_values: true,
-                    },
-                    EncodingType::Decimal {
-                        cell_at_least_one: false,
-                        cell_at_most_one: true,
-                        sudoku_has_all_values: true,
-                        sudoku_has_unique_values: false,
-                    },
-                    EncodingType::Decimal {
-                        cell_at_least_one: true,
-                        cell_at_most_one: true,
-                        sudoku_has_all_values: true,
-                        sudoku_has_unique_values: false,
-                    },
-                    EncodingType::Decimal {
-                        cell_at_least_one: false,
-                        cell_at_most_one: true,
-                        sudoku_has_all_values: true,
-                        sudoku_has_unique_values: true,
-                    },
-                    EncodingType::Decimal {
-                        cell_at_least_one: true,
-                        cell_at_most_one: true,
-                        sudoku_has_all_values: true,
-                        sudoku_has_unique_values: true,
-                    },
-                    EncodingType::Binary,
-                ];
-
-                let clues = self.get_option_value_sudoku();
-
-                if self.state.process_multithreaded {
-                    let dispatch_amount = match available_parallelism() {
-                        Ok(n) => n.get(),
-                        Err(_) => 2,
-                    };
-
-                    for chunk in encodings.iter().as_slice().chunks(dispatch_amount) {
-                        let mut handles = Vec::new();
-
-                        for encoding in chunk.iter().copied() {
-                            let clues = clues.clone();
-                            let history = self.state.history.clone();
-
-                            let handle = thread::spawn(move || {
-                                let mut solver = cadical::Solver::with_config("plain").unwrap();
-
-                                let res = solve_sudoku(&clues, &mut solver, &encoding);
-                                if let Ok(res) = res {
-                                    let cadical_stats = solver.stats();
-                                    let stats = Statistics::from_cadical_stats(
-                                        cadical_stats,
-                                        encoding,
-                                        clues,
-                                        res,
-                                    );
-
-                                    let mut history = history.lock().unwrap();
-                                    history.push(stats);
-                                }
-                            });
-
-                            handles.push(handle);
-                        }
-
-                        for handle in handles {
-                            handle.join().unwrap();
-                        }
-                    }
-                } else {
-                    for encoding in &encodings {
-                        let mut solver = cadical::Solver::with_config("plain").unwrap();
-
-                        let res = solve_sudoku(&clues, &mut solver, encoding);
-                        if let Ok(res) = res {
-                            let cadical_stats = solver.stats();
-                            let stats = Statistics::from_cadical_stats(
-                                cadical_stats,
-                                *encoding,
-                                clues.clone(),
-                                res,
-                            );
-
-                            let mut history = self.state.history.lock().unwrap();
-                            history.push(stats);
-                        }
-                    }
-                }
-
-                self.state.show_statistics = true;
-            }
-
-            ui.checkbox(
-                &mut self.state.process_multithreaded,
-                RichText::new("Parallel").size(text_scale),
-            );
-        });
-
-        if self.state.show_statistics {
-            ctx.show_viewport_immediate(
-                egui::ViewportId::from_hash_of("immediate_viewport_statistics"),
-                egui::ViewportBuilder::default()
-                    .with_title("Statistics")
-                    .with_inner_size([550.0, 275.0]),
-                |ctx, _class| {
-                    if ctx.input(|i| i.viewport().close_requested()) {
-                        self.state.show_statistics = false;
-                    }
-
-                    egui::CentralPanel::default().show(ctx, |ui| {
-                        let width = ui.available_width();
-                        let text_scale = (width / 60.0).max(10.0);
-
-                        ui.vertical(|ui| {
-                            ui.horizontal(|ui| {
-                                if ui.button("Clear history").clicked() {
-                                    let mut history = self.state.history.lock().unwrap();
-                                    history.clear();
-                                }
-                                if ui.button("Export as csv").clicked() {
-                                    self.export_as_csv();
-                                }
-                            });
-
-                            ScrollArea::vertical()
-                                .auto_shrink([false; 2])
-                                .stick_to_bottom(false)
-                                .show_viewport(ui, |ui, _viewport| {
-                                    let history = self.state.history.lock().unwrap();
-                                    for (i, his) in history.iter().rev().enumerate() {
-                                        ui.label(
-                                            RichText::new(format!("Sudoku {}:", i + 1))
-                                                .size(text_scale),
-                                        );
-
-                                        let old_spacing = ui.spacing().item_spacing;
-                                        ui.spacing_mut().item_spacing.y = 0f32;
-
-                                        ui.label(RichText::new("Clues:\tSolved:").size(text_scale));
-
-                                        for (clue_row, res_row) in
-                                            his.clues.iter().zip(his.sudoku.iter())
-                                        {
-                                            let mut chars: Vec<u8> = clue_row
-                                                .iter()
-                                                .map(|n| {
-                                                    if let Some(n) = n {
-                                                        *n as u8 + b'0'
-                                                    } else {
-                                                        b'X'
-                                                    }
-                                                })
-                                                .collect();
-                                            chars.push(b' ');
-                                            chars.extend(
-                                                res_row.iter().map(|n| n.unwrap() as u8 + b'0'),
-                                            );
-                                            let st = std::str::from_utf8(&chars).unwrap();
-                                            ui.label(RichText::new(st).size(text_scale / 1.5));
-                                            ui.spacing_mut().item_spacing.y = 0f32;
-                                        }
-
-                                        ui.spacing_mut().item_spacing = old_spacing;
-
-                                        ui.label(
-                                            RichText::new(format!("Time: {:.2}s", his.real_time))
-                                                .size(text_scale),
-                                        );
-                                        ui.label(
-                                            RichText::new(format!(
-                                                "Memory usage: {:.2}mb",
-                                                his.max_resident_set_size_mb
-                                            ))
-                                            .size(text_scale),
-                                        );
-                                        ui.label(
-                                            RichText::new(format!("Conflicts: {}", his.conflicts))
-                                                .size(text_scale),
-                                        );
-                                        ui.label(
-                                            RichText::new(format!(
-                                                "Learned clauses: {}",
-                                                his.learned_clauses
-                                            ))
-                                            .size(text_scale),
-                                        );
-                                        ui.label(
-                                            RichText::new(format!(
-                                                "Learned literals: {}",
-                                                his.learned_literals
-                                            ))
-                                            .size(text_scale),
-                                        );
-                                        ui.label(
-                                            RichText::new(format!("Decisions: {}", his.decisions))
-                                                .size(text_scale),
-                                        );
-                                        ui.label(
-                                            RichText::new(format!("Restarts: {}", his.restarts))
-                                                .size(text_scale),
-                                        );
-                                        ui.label(
-                                            RichText::new(format!(
-                                                "Encoding: {}",
-                                                match his.encoding {
-                                                    EncodingType::Binary => "Binary",
-                                                    EncodingType::Decimal { .. } => "Decimal",
-                                                }
-                                            ))
-                                            .size(text_scale),
-                                        );
-
-                                        match his.encoding {
-                                            EncodingType::Decimal {
-                                                cell_at_least_one,
-                                                cell_at_most_one,
-                                                sudoku_has_all_values,
-                                                sudoku_has_unique_values,
-                                            } => {
-                                                ui.label(
-                                                    RichText::new("Encoding rules:")
-                                                        .size(text_scale),
-                                                );
-                                                ui.label(
-                                                    RichText::new(format!(
-                                                        "Cell at least one: {}\n\
-                                                        Cell at most one: {}\n\
-                                                        Sudoku has all values: {}\n\
-                                                        Sudoku has unique values: {}",
-                                                        cell_at_least_one,
-                                                        cell_at_most_one,
-                                                        sudoku_has_all_values,
-                                                        sudoku_has_unique_values
-                                                    ))
-                                                    .size(text_scale / 1.5),
-                                                );
-                                            }
-                                            EncodingType::Binary => (),
-                                        }
-
-                                        ui.label(RichText::new("").size(text_scale));
-                                    }
-                                });
-                        });
-                    });
-                },
-            )
-        }
-    }
-
-    /// Save the statistics to a csv file
-    fn export_as_csv(&mut self) {
-        if let Some(file_path) = rfd::FileDialog::new()
-            .set_file_name("sudoku_statistics.csv")
-            .save_file()
-        {
-            let history = self.state.history.lock().unwrap();
-            let mut csv_string = String::new();
-            csv_string.push_str(&Statistics::csv_header());
-            for statistics in history.iter() {
-                csv_string.push_str(&statistics.csv());
-            }
-            let save_result = write_sudoku(csv_string, &file_path);
-            if let Err(e) = save_result {
-                self.current_error = Some(e);
-            }
-        }
     }
 
     /// Row for CNF encoding related inputs
